+++ src/AngleSharp.Css/Dom/Internal/CssRule.cs
@@ -1,5 +1,5 @@
 #nullable disable
-﻿namespace AngleSharp.Css.Dom
+namespace AngleSharp.Css.Dom
 {
     using AngleSharp.Css.Parser;
     using AngleSharp.Dom;