+++ src/AngleSharp.Css/Dom/Internal/Rules/CssDocumentRule.cs
@@ -72,8 +72,7 @@ namespace AngleSharp.Css.Dom
 
         public override void ToCss(TextWriter writer, IStyleFormatter formatter)
         {
-            var includeComments = formatter is ICommentPreservingFormatter commentFormatter && commentFormatter.PreserveComments;
-            var rules = formatter.BlockRules(GetFormattableRules(includeComments));
+            var rules = formatter.BlockRules(GetFormattableRules());
             writer.Write(formatter.Rule(RuleNames.Document, ConditionText, rules));
         }