+++ src/AngleSharp.Css/Dom/Internal/Rules/CssMediaRule.cs
@@ -51,8 +51,7 @@ namespace AngleSharp.Css.Dom
 
         public override void ToCss(TextWriter writer, IStyleFormatter formatter)
         {
-            var includeComments = formatter is ICommentPreservingFormatter commentFormatter && commentFormatter.PreserveComments;
-            var rules = formatter.BlockRules(GetFormattableRules(includeComments));
+            var rules = formatter.BlockRules(GetFormattableRules());
             writer.Write(formatter.Rule(RuleNames.Media, ConditionText, rules));
         }